Restructure enrichOffenceResultsPostPncEnrichment
