+++ src/lib/properties.ts
@@ -78,6 +78,71 @@ const MC_YOUTH = "MCY"
 const MC_ADULT = "MCA"
 const CROWN_COURT = "CC"
 
+const VICTIM_SURCHARGE_CREST_CODES = [
+  "COM",
+  "COMINST",
+  "COMTIME",
+  "FD",
+  "FDINST",
+  "FDTIME",
+  "FINE",
+  "PC",
+  "PCINST",
+  "PCTIME"
+]
+const VICTIM_SURCHARGE_AMOUNT_IN_POUNDS = 15
+const GUILTY_OF_ALTERNATIVE = "NA"
+const PNC_DISPOSAL_TYPE = {
+  VICTIM_SURCHARGE: "3117",
+  GUILTY_OF_ALTERNATIVE: "2060"
+}
+const RESULT_ADJOURNMENT_WITH_JUDGEMENT = "Adjournment with Judgement"
+const RESULT_JUDGEMENT_WITH_FINAL_RESULT = "Judgement with final result"
+const RESULT_UNRESULTED = "Unresulted"
+const RESULT_SENTENCE = "Sentence"
+const RESULT_ADJOURNMENT_POST_JUDGEMENT = "Adjournment post Judgement"
+const RESULT_ADJOURNMENT_PRE_JUDGEMENT = "Adjournment pre Judgement"
+const ADJOURNMENT_RANGES = [
+  [4001, 4009],
+  [4011, 4017],
+  [4020, 4021],
+  [4023, 4025],
+  [4027, 4035],
+  [4046, 4048],
+  [4050, 4050],
+  [4051, 4051],
+  [4053, 4058],
+  [4506, 4506],
+  [4508, 4508],
+  [4541, 4572],
+  [4574, 4574],
+  [4587, 4589]
+]
+const WARRANT_ISSUED_CODES = [[4575, 4577]]
+const ADJOURNMENT_NO_NEXT_HEARING_RANGES = [[0, 0]]
+const RESULT_CLASS_PLEAS = ["ADM"]
+const RESULT_CLASS_VERDICTS = ["NG", "NC", "NA"]
+const RESULT_CLASS_RESULT_CODES = [
+  // eslint-disable-next-line prettier/prettier
+  2050, 2063, 4010, 1016, 2053, 2060, 2065, 1029, 1030, 1044, 2006,
+  // eslint-disable-next-line prettier/prettier
+  3047, 3101, 3102, 3103, 3104, 3105, 3106, 3107, 3108, 3109, 3110,
+  // eslint-disable-next-line prettier/prettier
+  3111, 3126, 3127, 3128, 3129, 3130, 3131, 3146, 3147, 3148, 3272
+]
+const NON_RECORDABLE_RESULT_CODES = [
+  // eslint-disable-next-line prettier/prettier
+  1000, 1505, 1509, 1510, 1511, 1513, 1514, 2069, 2501, 2505, 2507,
+  // eslint-disable-next-line prettier/prettier
+  2508, 2509, 2511, 2514, 3501, 3502, 3503, 3504, 3508, 3509, 3510,
+  // eslint-disable-next-line prettier/prettier
+  3512, 3514, 4049, 4505, 4507, 4509, 4510, 4532, 4534, 4544, 4584,
+  // eslint-disable-next-line prettier/prettier
+  4585, 4586, 3118, 4592, 4593, 4594, 4595, 4596, 4597
+]
+const SUSPENDED_2ND_DURATION_RESULTS = [1115, 1134]
+const SUSPENDED = "Suspended"
+
 export {
   ENTERED_IN_ERROR_RESULT_CODE,
   STOP_LIST,
@@ -104,5 +169,24 @@ export {
   YOUTH_COURT,
   MC_YOUTH,
   MC_ADULT,
-  CROWN_COURT
+  CROWN_COURT,
+  VICTIM_SURCHARGE_CREST_CODES,
+  VICTIM_SURCHARGE_AMOUNT_IN_POUNDS,
+  GUILTY_OF_ALTERNATIVE,
+  PNC_DISPOSAL_TYPE,
+  RESULT_ADJOURNMENT_WITH_JUDGEMENT,
+  RESULT_JUDGEMENT_WITH_FINAL_RESULT,
+  RESULT_UNRESULTED,
+  RESULT_SENTENCE,
+  RESULT_ADJOURNMENT_POST_JUDGEMENT,
+  RESULT_ADJOURNMENT_PRE_JUDGEMENT,
+  ADJOURNMENT_RANGES,
+  WARRANT_ISSUED_CODES,
+  ADJOURNMENT_NO_NEXT_HEARING_RANGES,
+  RESULT_CLASS_PLEAS,
+  RESULT_CLASS_VERDICTS,
+  RESULT_CLASS_RESULT_CODES,
+  NON_RECORDABLE_RESULT_CODES,
+  SUSPENDED_2ND_DURATION_RESULTS,
+  SUSPENDED
 }
