+++ src/use-cases/enrichHearingOutcome/enrichFunctions/enrichOffenceResultsPostPncEnrichment/populateSourceOrganisation.ts
@@ -1,0 +1,33 @@
+import type { AnnotatedHearingOutcome, Result } from "../../../../types/AnnotatedHearingOutcome"
+import populateOrganisationUnitFields from "../../../populateOrganisationUnitFields"
+import { lookupOrganisationUnitByThirdLevelPsaCode } from "../../../dataLookup"
+
+const populateSourceOrganisation = (result: Result, hearingOutcome: AnnotatedHearingOutcome) => {
+  const { CourtHearingLocation, CourtHouseCode } = hearingOutcome.AnnotatedHearingOutcome.HearingOutcome.Hearing
+
+  if (result.SourceOrganisation) {
+    populateOrganisationUnitFields(result.SourceOrganisation)
+  }
+
+  if (!result.SourceOrganisation.OrganisationUnitCode && CourtHearingLocation) {
+    result.SourceOrganisation = { ...CourtHearingLocation }
+  }
+
+  if (!result.SourceOrganisation && CourtHouseCode) {
+    const organisationUnitData = lookupOrganisationUnitByThirdLevelPsaCode(CourtHouseCode)
+    if (organisationUnitData) {
+      const { topLevelCode, secondLevelCode, thirdLevelCode, bottomLevelCode } = organisationUnitData
+      result.SourceOrganisation = {
+        TopLevelCode: topLevelCode,
+        SecondLevelCode: secondLevelCode,
+        ThirdLevelCode: thirdLevelCode,
+        BottomLevelCode: bottomLevelCode,
+        OrganisationUnitCode: [topLevelCode, secondLevelCode, thirdLevelCode, bottomLevelCode].filter((x) => x).join()
+      }
+    }
+  }
+
+  populateOrganisationUnitFields(result.SourceOrganisation)
+}
+
+export default populateSourceOrganisation
